Fixed: * PIMS-890 Should get an error if you are trying to add more items to a shipment than are currently in inventory. * PIMS-905 Allow shipment items to be moved to more than one container at a time. * PIMS-1002 Improve error handling framework to allow services to validate domain objects and pass the validation errors back to the controller/gsp * PIMS-1006 Shipment Workflow > Child containers are not displayed [correctly] in the left nav of the Packing step. * PIMS-1003 Provide access to session-bound variables from within services. * Shipment Workflow > Refactored the saveItemAction into updateItemAction and saveItemAction.
git-svn-id: https://dev.pih-emr.org/svn/repository/projects/warehouse/trunk@10357 f0ac6d2b-dfc5-4988-9158-fc266d298d87

+++ grails-app/controllers/org/pih/warehouse/JsonController.groovy
@@ -31,7 +31,7 @@ class JsonController {
 		def quantity = 0
 		def warehouse = Warehouse.get(session.warehouse.id);
 		def lotNumber = (params.lotNumber) ? (params.lotNumber) : "";
-		def product = (params.productId) ? Product.get(params.productId) : null;
+		def product = (params.productId) ? Product.get(params.productId as Integer) : null;
 		
 		log.info "find by lotnumber '" + lotNumber + "' and product '" + product + "'";
 		def item = InventoryItem.findByLotNumberAndProduct(lotNumber, product)
